Add new CurrentInsuranceField to SE apartment form
