+++ apps/store/src/components/PriceForm/CurrentInsuranceField/CurrentInsuranceField.tsx
@@ -1,47 +1,77 @@
-import { useTranslation } from 'next-i18next'
-import { ChangeEventHandler, useState } from 'react'
-import { Space } from 'ui'
-import { InputSelect } from '@/components/InputSelect/InputSelect'
-import { InputSwitch } from '../InputSwitch'
-import { SelectOptions } from './CurrentInsuranceField.types'
+import { datadogLogs } from '@datadog/browser-logs'
+import { usePriceIntentDataUpdateMutation } from '@/services/apollo/generated'
+import { InputCurrentInsurance } from './InputCurrentInsurance'
 
-export type CurrentInsuranceFieldProps = {
+type Props = {
   label: string
-  companyOptions: SelectOptions
-  onCompanyChange: (company: string) => void
+  productName: string
+  priceIntentId: string
 }
 
-export const CurrentInsuranceField = (props: CurrentInsuranceFieldProps) => {
-  const { companyOptions, onCompanyChange, label } = props
-  const [hasInsurance, setHasInsurance] = useState(false)
-
-  const handleCheckedChange = (checked: boolean) => {
-    setHasInsurance(checked)
-  }
-
-  const { t } = useTranslation('purchase-form')
-
-  const handleChangeExternalInsurer: ChangeEventHandler<HTMLSelectElement> = (event) => {
-    if (event.target.value) {
-      onCompanyChange(event.target.value)
-    }
-  }
+export const CurrentInsuranceField = ({ label, productName, priceIntentId }: Props) => {
+  const companyOptions = useCompanyOptions(productName)
+  const handleCompanyChange = useUpdateExternalInsurer(priceIntentId)
 
   return (
-    <Space y={1}>
-      <InputSwitch label={label} checked={hasInsurance} onCheckedChange={handleCheckedChange} />
-
-      {hasInsurance && (
-        <InputSelect
-          name="externalInsurer"
-          // @TODO: remove, not part of updated designs
-          label="Current insurance company"
-          required
-          options={companyOptions}
-          placeholder={t('CURRENT_INSURANCE_FIELD_PLACEHOLDER')}
-          onChange={handleChangeExternalInsurer}
-        />
-      )}
-    </Space>
+    <InputCurrentInsurance
+      label={label}
+      companyOptions={companyOptions}
+      onCompanyChange={handleCompanyChange}
+    />
   )
+}
+
+const useCompanyOptions = (productName: string) => {
+  console.debug('Fetching companies for: ', productName)
+  const companyOptions = [
+    {
+      name: 'Trygg Hansa',
+      value: 'TRYGGHANSA',
+    },
+    {
+      name: 'Folksam',
+      value: 'FOLKSAM',
+    },
+    {
+      name: 'If',
+      value: 'IF',
+    },
+    {
+      name: 'Länsförsäkringar',
+      value: 'LANSFORSAKRINGAR',
+    },
+    {
+      name: 'Länsförsäkringar Stockholm',
+      value: 'LANSFORSAKRINGAR',
+    },
+    {
+      name: 'Moderna',
+      value: 'MODERNA',
+    },
+  ]
+  return companyOptions
+}
+
+const useUpdateExternalInsurer = (priceIntentId: string) => {
+  const [updateExternalInsurer] = usePriceIntentDataUpdateMutation({
+    onCompleted(data) {
+      if (data.priceIntentDataUpdate.priceIntent) {
+        console.log('Added external insurer')
+      } else {
+        datadogLogs.logger.warn('Failed to add external insurer', {
+          priceIntentId,
+        })
+      }
+    },
+    onError(error) {
+      datadogLogs.logger.warn('Error adding external insurer', {
+        priceIntentId,
+        error,
+      })
+    },
+  })
+
+  return (externalInsurerId: string) => {
+    updateExternalInsurer({ variables: { priceIntentId, data: { externalInsurerId } } })
+  }
 }
