--- apps/store/src/components/PriceForm/CurrentInsuranceField/CurrentInsuranceField.types.ts
@@ -1,1 +1,0 @@
-export type SelectOptions = ReadonlyArray<{ name: string; value: string }>
