+++ apps/store/src/services/PriceForm/data/SE_APARTMENT.ts
@@ -20,24 +20,6 @@ export const SE_APARTMENT: Template = {
       ],
     },
     {
-      id: 'insured-people',
-      title: { key: 'Insured people' },
-      submitLabel: { key: 'Next step' },
-      items: [
-        {
-          field: {
-            type: 'householdSize',
-            name: 'numberCoInsured',
-            label: { key: 'Household size' },
-            required: true,
-            defaultValue: 0,
-            max: 5,
-          },
-          layout: { columnSpan: 6 },
-        },
-      ],
-    },
-    {
       id: 'your-home',
       title: { key: 'Your home' },
       submitLabel: { key: 'Calculate price' },
@@ -91,6 +73,32 @@ export const SE_APARTMENT: Template = {
             min: 0,
           },
           layout: { columnSpan: 3 },
+        },
+        {
+          field: {
+            type: 'current-insurance',
+            name: 'externalInsurer',
+            label: { key: 'I already have home insurance' },
+          },
+          layout: { columnSpan: 6 },
+        },
+      ],
+    },
+    {
+      id: 'insured-people',
+      title: { key: 'Insured people' },
+      submitLabel: { key: 'Next step' },
+      items: [
+        {
+          field: {
+            type: 'householdSize',
+            name: 'numberCoInsured',
+            label: { key: 'Household size' },
+            required: true,
+            defaultValue: 0,
+            max: 5,
+          },
+          layout: { columnSpan: 6 },
         },
       ],
     },
